### 2018-01-09 #529 @whedon
[PRE REVIEW]: pyneqsys: Solve symbolically defined systems of non-linear equations numerically
**Submitting author:** @bjodah (<a href="http://orcid.org/0000-0003-0596-0222">Björn Dahlgren</a>)
**Repository:** <a href="https://github.com/bjodah/pyneqsys" target ="_blank">https://github.com/bjodah/pyneqsys</a>
**Version:** v0.5.1
