### 2018-01-09 #529 @bjodah
perhaps @arghdos would be willing to review this paper? `pyneqsys` is similar in structure to the earlier submitted pyodesys package.
