### 2018-01-10 #529 @bjodah
Or in the interest of spreading the work on reviewers, maybe @oesteban is interested in reviewing?
