### 2018-01-11 #529 @arghdos
@bjodah I'm going to be a bit crazy for the next few weeks trying to wrap up a paper, so it might go faster if someone else reviews.  I can do it if you don't find another reviewer though

-- arghdos
