### 2018-01-11 #529 @bjodah
@arghdos perfectly understandable‒good luck with the paper! (and thanks for the offer still)

Judging by what I see on Andreas Mayer's profile (@andim) I think it could be possible that he'd be willing to review?
