### 2018-01-16 #529 @arfon
> @arfon have I understood it correctly that only an editor can set reviwer using whedon?

Yep. I'll go ahead and do that now.
