### 2018-01-17 bjodah/pyneqsys#33 @andim
Adding DOIs to the references
[This issue is part of the JOSS review, see openjournals/joss-reviews/issues/531.]

Could you please add DOIs to the references where not already done? At least for the sundials paper their appears to be one (10.1145/1089014.1089020) potentially also for some of the other references.
