### 2018-01-17 5e89a86
Update README, add DOI & URN to JOSS references
