+++ joss-paper/paper.bib
@@ -12,6 +12,8 @@
 }
 
 @article{hindmarsh2005sundials,
+  doi = {10.1145/1089014.1089020},
+  url = {https://doi.org/10.1145/1089014.1089020},
   title={SUNDIALS: Suite of nonlinear and differential/algebraic equation solvers},
   author={Hindmarsh, Alan C and Brown, Peter N and Grant, Keith E and Lee, Steven L and Serban, Radu and Shumaker, Dan E and Woodward, Carol S},
   journal={ACM Transactions on Mathematical Software (TOMS)},
@@ -42,7 +44,8 @@
   title={A Family of Newton Codes for Systems of Highly Nonlinear Equations},
   author={Weimann, UN},
   year={1991},
-  institution={ZIB Technical Report TR-91-10, ZIB, Berlin, Germany}
+  institution={ZIB Technical Report TR-91-10, ZIB, Berlin, Germany},
+  note={urn:nbn:de:0297-zib-4855}
 }
 
 @manual{mpmath,
