### 2018-01-17 b72ab0e
Check README using rstcheck
