### 2018-01-17 6e16025
About to release v0.5.2
