### 2018-01-17 bjodah/pyneqsys#33 @bjodah
Thanks for the SUNDIALS doi (I missed it the first time). For the other references I came up empty handed. However the NLEQ2 now has a "urn" identifier which [ZIB](https://opus4.kobv.de/opus4-zib/frontdoor/index/index/docId/485) asked that citing parties use.
