### 2018-01-17 bjodah/pyneqsys#33 @andim
Sounds good. I don't want to send you hunting DOIs that don't exist, but you could add the one for Sundials. :+1:
