### 2018-01-17 bjodah/pyneqsys#33 @bjodah
Thanks, I added it: https://github.com/bjodah/pyneqsys/blob/master/joss-paper/paper.bib#L15
