### 2018-01-18 bjodah/pyneqsys#34 @andim
Improve docstrings of subclasses
[This issue is part of the JOSS review, see openjournals/joss-reviews#531.]

The package makes heavy use of subclasses, but most of the documentation is given on the level of the base classes only. It would be great to expose more of that documentation directly on the level of the user-facing class. This would be helpful for the online documentation (as it would no longer require clicking through the class hierarchy to get to the documentation) and it would be crucial for interactive work. For example calling `help(SymbolicSys.from_callback)` gives only
```
from_callback(cb, nx=None, nparams=None, **kwargs) method of builtins.type instance
    Generate a SymbolicSys instance from a callback
```
which does not help the user in knowing how the callback function should be constructed. To avoid duplication you can of course simply reuse the docstring of the baseclass, where applicable. I believe there is also some automated ways of solving this issue of docstring inheritance using decorators.
