### 2018-01-18 bjodah/pyneqsys#35 @andim
Add short description of problem to be solved to examples
[This issue is part of the JOSS review, see openjournals/joss-reviews#531.]

The examples are very good to document the functionality of the software and especially the notebooks are well-documented. For the pure python examples it would be great to have a short description of the problem to be solved as a comment in the file.
