### 2018-01-19 bd32591
Improving documentation throughout
