### 2018-01-20 10310a5
Some suggestions for the paper
I allow myself to propose a few edits to the text to streamline its flow and clarify some things. Feel free to pick those that you like and disregard the others.
