### 2018-01-21 efcab5a
Enhanced docstrings
