+++ pyneqsys/core.py
@@ -91,7 +91,7 @@ class _NeqSysBase(object):
         params : array_like
             Parameter values
         vaired_data : array_like
-            numerical values of the varied parameter
+            Numerical Values Of The varied parameter.
         varied_idx : int or str
             Index of the varied parameter (indexing starts at 0).
             If ``self.par_by_name`` this should be the name (str) of the varied
@@ -162,7 +162,15 @@ class _NeqSysBase(object):
 
         Parameters
         ----------
-        xres :
+        xres : array
+            Of shape ``(varied_data.size, self.nx)``.
+        varied_data : array
+            See :meth:`solve_series`.
+        varied_idx : int or str
+            See :meth:`solve_series`.
+        \\*\\*kwargs :
+            Keyword arguments passed to :func:`pyneqsys.plotting.plot_series`.
+
         """
         for attr in 'names latex_names'.split():
             if kwargs.get(attr, None) is None:
@@ -177,6 +185,7 @@ class _NeqSysBase(object):
         return ax
 
     def plot_series_residuals(self, xres, varied_data, varied_idx, params, **kwargs):
+        """ Analogous to :meth:`plot_series` but will plot residuals. """
         nf = len(self.f_cb(*self.pre_process(xres[0], params)))
         xerr = np.empty((xres.shape[0], nf))
         new_params = np.array(params)
@@ -187,6 +196,7 @@ class _NeqSysBase(object):
         return self.plot_series(xerr, varied_data, varied_idx, **kwargs)
 
     def plot_series_residuals_internal(self, varied_data, varied_idx, **kwargs):
+        """ Analogous to :meth:`plot_series` but for internal residuals from last run. """
         nf = len(self.f_cb(*self.pre_process(
             self.internal_xout[0], self.internal_params_out[0])))
         xerr = np.empty((self.internal_xout.shape[0], nf))
@@ -196,7 +206,11 @@ class _NeqSysBase(object):
 
     def solve_and_plot_series(self, x0, params, varied_data, varied_idx, solver=None, plot_kwargs=None,
                               plot_residuals_kwargs=None, **kwargs):
-        """ Solve and plot for a series of a varied parameter """
+        """ Solve and plot for a series of a varied parameter.
+
+        Convenience method, see :meth:`solve_series`, :meth:`plot_series` &
+        :meth:`plot_series_residuals_internal` for more information.
+        """
         sol, nfo = self.solve_series(
             x0, params, varied_data, varied_idx, solver=solver, **kwargs)
         ax_sol = self.plot_series(sol, varied_data, varied_idx, info=nfo,
@@ -212,14 +226,15 @@ class _NeqSysBase(object):
 
 
 class NeqSys(_NeqSysBase):
-    """Represent a system of non-linear equations
+    """Represents a system of non-linear equations.
 
-    Object representing nonlinear equation system.
-    Provides unified interface to:
+    This class provides a unified interface to:
 
     - scipy.optimize.root
     - NLEQ2
     - KINSOL
+    - mpmath
+    - levmar
 
     Parameters
     ----------
@@ -238,6 +253,16 @@ class NeqSys(_NeqSysBase):
         Names of variables, used for plotting and for referencing by name.
     param_names : iterable of strings (default: None)
         Names of the parameters, used for referencing parameters by name.
+    x_by_name : bool, default: ``False``
+        Will values for *x* be referred to by name (in dictionaries)
+        instead of by index (in arrays)?
+    par_by_name : bool, default: ``False``
+        Will values for parameters be referred to by name (in dictionaries)
+        instead of by index (in arrays)?
+    latex_names : iterable of str, optional
+        Names of variables in LaTeX format.
[…]
