### 2018-01-21 6ae7fef
Add docs to .plotting
