### 2018-01-21 bjodah/pyneqsys#34 @bjodah
When I went over them once more I can nothing but aggree. I tried to address this in gh-36, what do you think about it?
