### 2018-01-21 bjodah/pyneqsys#35 @bjodah
Glad to hear you like what you saw in the notebooks.
The examples were, as you noted, very terse. I tried to address this in gh-36 as well, let me know what you think.
