### 2018-01-21 bjodah/pyneqsys#34 @andim
That looks like a good fix to the issue ! Thanks for putting that much effort into this - I am sure it will pay off for future users.
