### 2018-01-21 bjodah/pyneqsys#35 @andim
Looks good!
