### 2018-01-21 #531 @andim
It has been a pleasure to work with @bjodah to review his submission of  `pyneqsys` to JOSS over this past week. All the issues I've raised (linked in this review thread above) have been addressed satisfactorily and some additional feedback incorporated as well (PR https://github.com/bjodah/pyneqsys/pull/31, https://github.com/bjodah/pyneqsys/pull/32, https://github.com/bjodah/pyneqsys/pull/37).

What are the next steps for acceptance of the paper @arfon?
