### 2018-01-22 3a4f33f
Build conda recipe for for 2.7 & 3.6
