### 2018-01-22 #531 @bjodah
Thank you @arfon, and thanks again @andim
The JOSS's submission process is really amazing.
